Fix T89487: Crash adding Rigid Body to object with shared mesh data
Not sure why this bug was only discovered by such an elaborate steps
and why it took so long to be discovered. The root of the issue is
that in the 956c539e597a the typical flow of tag+flush+evaluate was
violated and tagging for visibility change happened after flush.